Separated main methods and App widget
